movie_library/requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
django-phonenumbers==1.0.1
djangorestframework==3.15.2
phonenumbers==8.13.49
psycopg2==2.9.10
sqlparse==0.5.1
